Move base button
